requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
